renamed the behaviour and corrected the typo
