+++ apps/controlled-drugs/data/trading-reasons.json
@@ -129,7 +129,7 @@
   },
   {
     "label": "Substance misuse service",
-    "value": "substabce-misuse-service"
+    "value": "substance-misuse-service"
   },
   {
     "label": "Transport and distribution services",
